Attempted to implement iPart (failure?)
iParts don't seem to do all the things I need for the variations. This
is a pretty small project anyway, so the added complexity just isn't
worth it. I already have a spreadsheet set up, so parameter changes
shouldn't be too hard. iParts cause problems with materials and the
thumb keycap.

Calling it a failure and leaving it here. Maybe I'll come back? I doubt
it.