Added pole height, shortened keycap
To prevent collisions between the keycap skirt and the keyswitch, the
actual keycap section was foreshortened. This shouldn't really change
the aesthetic of the keycaps, because they're so tall, but that's the
least of our concerns.

The wall thickness of the pole is 0.4mm too thin - we'll see if this can
be rectified without causing clearance issues with the keyswitch. If
not, we may need to print forks instead of a pole - which will be
equally interesting.

Also, found a bug in Inventor that meant I had to set the capOffset of
the thumb variant to 0.001 instead of 0 - otherwise, direction was not
preserved (a big problem).