Major change to Cherry MX pole design
The new design should work (in theory) *and* be much easier to 3D print.
The traditional design has walls that are far too thin - not ideal.

Also uprgaded the ribbing to have a slight draft angle, for strength
reasons.

NOTE: The design may also change in the future to incorporate walls on
the long side of the keyswitch. This could potentially improve the
strength of the model, but we'll see how it holds up without. Not adding
until a test print because it could cause problems with clearances on
the keyswitch.