Verified Dimensions against physical switches
Very slight adjustments were made to the dimensions that I believe will
improve the fit of the keycap to the switch. Also, the cut-in depth was
reduced, which will increase the strength with marginal extra print
time.

The project is almost complete - all that's left is to do test prints
and mark it off as finished!